+++ crates/core/src/wrappers/controllable.rs
@@ -45,11 +45,11 @@ where S: Sound
 impl<S> Sound for Controllable<S>
 where S: Sound
 {
-    fn channel_count(&self) -> u16 {
+    fn channel_count(&self) -> usize {
         self.inner.channel_count()
     }
 
-    fn sample_rate(&self) -> u32 {
+    fn sample_rate(&self) -> usize {
         self.inner.sample_rate()
     }
 
@@ -67,7 +67,7 @@ where S: Sound
         self.inner.on_start_of_batch();
     }
 
-    fn fill_next_frames(&mut self, buffer: &mut [i16]) -> Result<(usize, NextState), RawedioError> {
+    fn fill_next_frames(&mut self, buffer: &mut [f32]) -> Result<(usize, NextState), RawedioError> {
         let next = self.inner.fill_next_frames(buffer)?;
         match next {
             (_, NextState::Playing | NextState::MetadataChanged | NextState::Paused) => Ok(next),
